+++ server/src/models/iface.go
@@ -1,0 +1,1 @@
+package models
